% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp_scrape_web.R
\name{.annotate_site}
\alias{.annotate_site}
\title{Annotate Website Data}
\usage{
.annotate_site(site)
}
\arguments{
\item{site}{A data.table representing the scraped website data.}
}
\value{
A data.table with annotations.
}
\description{
This function processes a data.table containing website data and annotates
each line, identifying relevant text and removing boilerplate content.
}
\keyword{internal}
