% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_search.R
\name{.decode_duckduckgo_urls}
\alias{.decode_duckduckgo_urls}
\title{Decode DuckDuckGo Redirect URLs}
\usage{
.decode_duckduckgo_urls(redirected_urls)
}
\arguments{
\item{redirected_urls}{A vector of DuckDuckGo search result URLs.}
}
\value{
A vector of decoded URLs.
}
\description{
This function decodes the DuckDuckGo search result URLs that are redirected.
}
