% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_search.R
\name{.extract_links}
\alias{.extract_links}
\title{Extract links from a search engine result page}
\usage{
.extract_links(search_url)
}
\arguments{
\item{search_url}{The URL of the search engine result page.}
}
\value{
A character vector of URLs.
}
\description{
This function extracts all the links (href attributes) from a search engine result page.
}
