% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.generate_random_ids}
\alias{.generate_random_ids}
\title{Generate Random IDs}
\usage{
.generate_random_ids(n)
}
\arguments{
\item{n}{The number of IDs to generate.}
}
\value{
A vector of random IDs.
}
\description{
This function generates a specified number of random IDs.
}
