% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.is_valid_json}
\alias{.is_valid_json}
\title{Validate JSON String}
\usage{
.is_valid_json(json_string)
}
\arguments{
\item{json_string}{The string to validate.}
}
\value{
TRUE if the string is valid JSON, FALSE otherwise.
}
\description{
This function checks if a given string is a valid JSON.
}
