% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_api_completions.R
\name{.openai_chat_completions}
\alias{.openai_chat_completions}
\title{OpenAI Chat Completions}
\usage{
.openai_chat_completions(
  model = "gpt-3.5-turbo",
  system_message = "",
  user_message = "",
  temperature = 1,
  top_p = 1,
  max_tokens = NULL,
  openai_api_key,
  openai_organization,
  is_json_output = TRUE
)
}
\arguments{
\item{model}{The model to use for the API call.}

\item{system_message}{The message provided by the system (e.g., instructions or context).}

\item{user_message}{The message provided by the user.}

\item{temperature}{The temperature for the model's output.}

\item{top_p}{The top-p sampling value.}

\item{max_tokens}{The maximum number of tokens to generate.}

\item{openai_api_key}{The API key for the OpenAI API.}

\item{openai_organization}{The organization ID for the OpenAI API.}

\item{is_json_output}{A logical indicating whether the output should be JSON.}

\item{max_attempts}{The maximum number of attempts to make for generating valid output.}
}
\value{
The generated text.
}
\description{
This function interacts with the OpenAI API to generate text completions.
}
