% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rag_get_openai_embs.R
\name{.openai_embs}
\alias{.openai_embs}
\title{Internal function for fetching embeddings}
\usage{
.openai_embs(x)
}
\arguments{
\item{x}{A character string or vector of texts to fetch embeddings for.}
}
\description{
Internal function for fetching embeddings
}
\keyword{internal}
