% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_search.R
\name{.process_duckduckgo}
\alias{.process_duckduckgo}
\title{Process DuckDuckGo search results}
\usage{
.process_duckduckgo(
  search_term,
  num_pages,
  time_filter,
  insite,
  intitle,
  combined_pattern
)
}
\arguments{
\item{search_term}{The search query.}

\item{num_pages}{Number of result pages to retrieve.}

\item{time_filter}{Optional time filter ("week", "month", "year").}

\item{insite}{Restrict search to a specific domain.}

\item{intitle}{Search within the title.}

\item{combined_pattern}{A pattern for filtering out irrelevant URLs.}
}
\value{
A `data.table` of search results from DuckDuckGo.
}
\description{
This function handles the extraction of search results from DuckDuckGo.
}
