% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_search.R
\name{.process_yahoo}
\alias{.process_yahoo}
\title{Process Yahoo News search results}
\usage{
.process_yahoo(search_term, num_pages, combined_pattern = combined_pattern)
}
\arguments{
\item{search_term}{The search query.}

\item{num_pages}{Number of result pages to retrieve.}

\item{combined_pattern}{A pattern for filtering out irrelevant URLs.}
}
\value{
A `data.table` of search results from Yahoo News.
}
\description{
This function retrieves and processes search results from Yahoo News,
automatically sorting by the most recent articles.
}
