% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_search_corpus.R
\name{.translate_query}
\alias{.translate_query}
\title{Translate Search Query}
\usage{
.translate_query(x)
}
\arguments{
\item{x}{The search query string to be translated.}
}
\value{
A character string representing the translated query.
}
\description{
Translates a search query into a format suitable for regex matching,
particularly for inline searches.
}
\keyword{internal}
