% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emb_call_huggingface.R
\name{emb_call_huggingface}
\alias{emb_call_huggingface}
\title{Call Hugging Face API for Embeddings}
\usage{
emb_call_huggingface(
  tif,
  text_hierarchy,
  api_token,
  api_url =
    "https://api-inference.huggingface.co/pipeline/feature-extraction/sentence-transformers/all-MiniLM-L6-v2",
  query = NULL,
  dims = 384,
  batch_size = 250,
  sleep_duration = 1
)
}
\arguments{
\item{tif}{A data frame containing text data.}

\item{text_hierarchy}{A character vector indicating the columns used to create row names.}

\item{api_token}{Token for accessing the Hugging Face API.}

\item{api_url}{The URL of the Hugging Face API endpoint (default is set to a specific model endpoint).}

\item{query}{An optional single text query for which embeddings are required.}

\item{dims}{The dimension of the output embeddings.}

\item{batch_size}{Number of rows in each batch sent to the API.}

\item{sleep_duration}{Duration in seconds to pause between processing batches.}
}
\value{
A matrix containing embeddings, with each row corresponding to a text input.
}
\description{
Retrieves embeddings for text data using Hugging Face's API. It can process a batch of texts or a single query.  Mostly for demo purposes.
}
