% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_get_neighbors.R
\name{get_sim}
\alias{get_sim}
\title{Compute Similarity Between Matrices}
\usage{
get_sim(x, y = NULL, norm = c("l2", "none"))
}
\arguments{
\item{x}{A numeric matrix or a sparse matrix.}

\item{y}{An optional numeric matrix or a sparse matrix to be compared with \code{x}.}

\item{norm}{Character, either "l2" or "none" for normalization method.}
}
\value{
A matrix of cosine similarity values.
}
\description{
This function computes the cosine similarity between two matrices,
normalizing them based on the specified norm.
}
\examples{
x <- matrix(rnorm(50), nrow = 10)
# Self-similarity
sim_matrix <- get_sim(x)
# Similarity with another matrix
y <- matrix(rnorm(50), nrow = 10)
sim_matrix <- get_sim(x, y)
}
