% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_api_completions.R
\name{llm_api_completions}
\alias{llm_api_completions}
\title{LLM API Completions}
\usage{
llm_api_completions(
  id,
  user_message = "",
  annotators = 1,
  model_name = "gpt-3.5-turbo",
  temperature = 1,
  top_p = 1,
  max_tokens = NULL,
  is_json_output = TRUE,
  max_attempts = 10,
  openai_api_key = Sys.getenv("OPENAI_API_KEY"),
  openai_organization = NULL,
  cores = 1
)
}
\arguments{
\item{id}{A unique identifier for the request.}

\item{user_message}{The message provided by the user.}

\item{annotators}{The number of annotators (default is 1).}

\item{model_name}{The name of the OpenAI model to use (default is 'gpt-3.5-turbo').}

\item{temperature}{The temperature for the model's output (default is 1).}

\item{top_p}{The top-p sampling value (default is 1).}

\item{max_tokens}{The maximum number of tokens to generate (default is NULL).}

\item{is_json_output}{A logical indicating whether the output should be JSON (default is TRUE).}

\item{max_attempts}{The maximum number of attempts to make for generating valid output (default is 10).}

\item{openai_api_key}{The API key for the OpenAI API (default is retrieved from environment variables).}

\item{openai_organization}{The organization ID for the OpenAI API (default is NULL).}

\item{cores}{The number of cores to use for parallel processing (default is 1).}
}
\value{
A data.table containing the generated text and metadata.
}
\description{
This function generates text using the OpenAI API.
}
\examples{
\dontrun{
llm_api_completions(id = "example_id", user_message = "What is the capital of France?", openai_api_key = "your_api_key")
}
}
