% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_get_openai_embs.R
\name{llm_get_openai_embs}
\alias{llm_get_openai_embs}
\title{Extract Embeddings from OpenAI}
\usage{
llm_get_openai_embs(
  tif,
  batch.id,
  text.segment,
  text.segment.id,
  query = NULL,
  wait = 30
)
}
\arguments{
\item{x}{A character string representing the text to be embedded.}
}
\value{
A numeric vector representing the embeddings of the input text.
}
\description{
This function sends a text to the OpenAI API and retrieves its embeddings.
}
\examples{
# Example usage
# Sys.setenv(OPENAI_API_KEY = "your_api_key_here")
# wd_extract_openai_embs("Example text")
}
