% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp_cast_tokens.R
\name{nlp_cast_tokens}
\alias{nlp_cast_tokens}
\title{Convert Token List to Data Frame}
\usage{
nlp_cast_tokens(tok)
}
\arguments{
\item{tok}{A list where each element contains tokens corresponding to a document or a sentence.}
}
\value{
A data frame with columns for token name and token.
}
\description{
This function converts a list of tokens into a data frame, extracting and separating document and sentence identifiers if needed.
}
\examples{
tokens <- list(c("Hello", "world", "."),
               c("This", "is", "an", "example", "." ),
               c("This", "is", "a", "party", "!"))
names(tokens) <- c('1.1', '1.2', '2.1')
dtm <- nlp_cast_tokens(tokens)

}
