% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp_get_openai_embs.R
\name{nlp_fetch_openai_embs}
\alias{nlp_fetch_openai_embs}
\title{Extract Embeddings from OpenAI}
\usage{
nlp_fetch_openai_embs(
  tif,
  batch_id = "batch_id",
  text = "text",
  text_id = "text_id",
  query = NULL,
  wait = 30
)
}
\arguments{
\item{x}{A character string representing the text to be embedded.}
}
\value{
A numeric vector representing the embeddings of the input text.
}
\description{
This function sends a text to the OpenAI API and retrieves its embeddings.
}
