% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp_melt_tokens.R
\name{nlp_melt_tokens}
\alias{nlp_melt_tokens}
\title{Tokenize Data Frame by Specified Column(s)}
\usage{
nlp_melt_tokens(df, token, by = c("doc_id"))
}
\arguments{
\item{df}{A data frame containing the data to be tokenized.}

\item{token}{The name of the column in \code{df} that contains the tokens.}

\item{by}{A character vector indicating the column(s) by which to group the data.}
}
\value{
A list of vectors, each containing the tokens of a group defined by the \code{by} parameter.
}
\description{
This function tokenizes a data frame based on a specified token column and groups the data by one or more specified columns.
}
\examples{
# Example usage
# df <- data.frame(doc_id = c(1, 1, 2, 2), text = c("token1", "token2", "token3", "token4"))
nlp_df_token(df, "text", by = "doc_id")

}
