% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp_pretty_text.R
\name{nlp_pretty_text}
\alias{nlp_pretty_text}
\title{Pretty Print Text for NLP}
\usage{
nlp_pretty_text(x, width = 50L, char_length = NULL)
}
\arguments{
\item{x}{A character vector; the text to be formatted.}

\item{width}{An integer; the width to wrap the text at. Defaults to 50.}

\item{char_length}{An optional integer; the maximum character length of the text.
If specified, text longer than this will be truncated.}
}
\value{
The function prints formatted text and does not return a value.
}
\description{
Formats and prints text to make it more readable, especially useful in NLP tasks.
It allows the option to truncate text to a specific character length and to wrap
text to a specified width.
}
\examples{
nlp_pretty_text("This is an example of a longer string that will be wrapped for better readability", width = 40)
# With character length limit
nlp_pretty_text("This is an example of a longer string", width = 40, char_length = 20)

}
