% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp_search_annotation.R
\name{nlp_search_annotation}
\alias{nlp_search_annotation}
\title{Search in Document-Term Matrix}
\usage{
nlp_search_annotation(
  df,
  search_col,
  id_col,
  include,
  logic = "or",
  exclude = NULL
)
}
\arguments{
\item{df}{A Document-Term Matrix (df) or data frame to be searched.}

\item{search_col}{The name of the column in the df to search through.}

\item{id_col}{The name of the column in the df representing document IDs.}

\item{include}{A vector of terms to include in the search.}

\item{logic}{A character string specifying the logic to use: 'or' or 'and'.}

\item{exclude}{An optional vector of terms to exclude from the search.}
}
\value{
A subset of the df based on the search criteria.
}
\description{
This function searches within a Document-Term Matrix (df) using specified
inclusion and exclusion criteria and returns a subset of the df based on the search.
}
\examples{
df <- data.frame(doc_id = 1:3, text = c("apple banana", "banana cherry", "cherry apple"))
search_df(df, "text", "doc_id", include = c("apple", "banana"), logic = "or")
search_df(df, "text", "doc_id", include = c("apple", "banana"), logic = "and", exclude = "cherry")
}
