% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp_token_mwe.R
\name{nlp_token_mwe}
\alias{nlp_token_mwe}
\title{Convert Tokens to Multi-Word Expressions}
\usage{
nlp_token_mwe(tok, mwe, concat = "_")
}
\arguments{
\item{tok}{A list where each element contains tokens corresponding to a document or a sentence.}

\item{mwe}{Character vector of multi-word expressions to be identified and concatenated.}

\item{concat}{The string to use for concatenating the tokens in each multi-word expression.}
}
\value{
A list of tokens with specified multi-word expressions concatenated.
}
\description{
This function processes a list of tokens, identifying and concatenating multi-word expressions (MWEs)
based on specified patterns using the 'quanteda' package.
}
\examples{
# Example usage
# tok <- list(doc1 = c("the", "quick", "brown", "fox"), doc2 = c("jumps", "over", "the", "lazy", "dog"))
# mwe <- c("quick brown fox", "lazy dog")
# token2mwe(tok, mwe)

}
