% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlplite-package.R
\docType{package}
\name{nlpx-package}
\alias{nlpx}
\alias{nlpx-package}
\title{nlpx: Lightweight and User-Friendly Natural Language Processing Toolkit}
\description{
Provides a suite of tools for basic natural language processing tasks. Designed to be lightweight and user-friendly, it includes functionalities for text tokenization, basic text analysis, and pattern searching within text corpora. It's particularly suitable for quick analyses and for users who prefer a simple toolkit for NLP tasks.
}
\author{
\strong{Maintainer}: Jason Timm \email{first.last@example.com}

}
\keyword{internal}
