% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rag_batch_wordcount.R
\name{rag_batch_wordcount}
\alias{rag_batch_wordcount}
\title{Group Numeric Vector into Batches Based on Cumulative Sum}
\usage{
rag_batch_wordcount(x, threshold)
}
\arguments{
\item{x}{A numeric vector to be batched.}

\item{threshold}{A numeric threshold for the cumulative sum of each batch.}
}
\value{
A numeric vector indicating the batch number for each element of \code{x}.
}
\description{
This function takes a numeric vector and groups its elements into batches.
Each batch's cumulative sum does not exceed a specified threshold.
}
