% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rag_build_chunks.R
\name{rag_build_chunks}
\alias{rag_build_chunks}
\title{Chunk Sentences for NLP}
\usage{
rag_build_chunks(df, chunk_level, chunk_size, context_size)
}
\arguments{
\item{df}{A data.table object containing the text data.}

\item{chunk_level}{The level at which to chunk the text: "sentence" or "paragraph"}

\item{chunk_size}{The number of sentences in each chunk.}

\item{context_size}{The number of sentences around each chunk to include as context.}
}
\value{
A data.table object with chunks of text and their context.
}
\description{
This function processes text data for NLP tasks by dividing text into chunks
based on sentences or paragraphs, and includes additional context around each chunk.
}
