% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rag_get_openai_embs.R
\name{rag_fetch_openai_embs}
\alias{rag_fetch_openai_embs}
\title{Fetch OpenAI Embeddings}
\usage{
rag_fetch_openai_embs(
  tif,
  batch_id = "batch_id",
  text = "text",
  text_id = "text_id",
  query = NULL,
  wait = 30
)
}
\arguments{
\item{tif}{A dataframe containing text data.}

\item{batch_id}{The name of the column in \code{tif} identifying each batch. Default is 'batch_id'.}

\item{text}{The name of the column in \code{tif} containing the text data. Default is 'text'.}

\item{text_id}{The name of the column in \code{tif} for text identifiers. Default is 'text_id'.}

\item{query}{An optional single text query for fetching embeddings.}

\item{wait}{Time in seconds to wait between processing batches. Default is 30 seconds.}
}
\value{
A matrix of embeddings for either a single query or a combined matrix for all batches.
}
\description{
Retrieves embeddings from OpenAI for a given text or a batch of texts.
It can process a single query or batches of text data from a dataframe.
}
