% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_add_context.R
\name{add_context}
\alias{add_context}
\alias{search_add_context}
\title{Add context to search results}
\usage{
search_add_context(gramx, df, inline = "inline", highlight = NULL)
}
\arguments{
\item{gramx}{output from find_gramx()}

\item{df}{Annotion dataframe}

\item{highlight}{A boolean}

\item{form}{A character string}

\item{tag}{A character string}
}
\value{
A data frame
}
\description{
Add context to search results
}
