% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_dtm.R
\name{search_dtm}
\alias{search_dtm}
\title{Search in Document-Term Matrix}
\usage{
search_dtm(dtm, search_col, id_col, include, logic = "or", exclude = NULL)
}
\arguments{
\item{dtm}{A Document-Term Matrix (DTM) or data frame to be searched.}

\item{search_col}{The name of the column in the DTM to search through.}

\item{id_col}{The name of the column in the DTM representing document IDs.}

\item{include}{A vector of terms to include in the search.}

\item{logic}{A character string specifying the logic to use: 'or' or 'and'.}

\item{exclude}{An optional vector of terms to exclude from the search.}
}
\value{
A subset of the DTM based on the search criteria.
}
\description{
This function searches within a Document-Term Matrix (DTM) using specified
inclusion and exclusion criteria and returns a subset of the DTM based on the search.
}
\examples{
dtm <- data.frame(doc_id = 1:3, text = c("apple banana", "banana cherry", "cherry apple"))
search_dtm(dtm, "text", "doc_id", include = c("apple", "banana"), logic = "or")
search_dtm(dtm, "text", "doc_id", include = c("apple", "banana"), logic = "and", exclude = "cherry")
}
