% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_regex.R
\name{search_regex}
\alias{search_regex}
\title{Search corpus via regex}
\usage{
search_regex(corpus, query, by, highlight = c("<b>", "</b>"))
}
\arguments{
\item{corpus}{A data frame or data.table with a \code{text} column.}

\item{query}{The search pattern (regex).}

\item{by}{Character vector of identifier columns.}

\item{highlight}{Length-two character vector (default \code{c('<b>', '</b>')}).}
}
\description{
Search corpus via regex
}
