% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_semantics.R
\name{search_semantics}
\alias{search_semantics}
\title{Find Nearest Neighbors Based on Cosine Similarity}
\usage{
search_semantics(x, matrix, n = 10)
}
\arguments{
\item{x}{A character or numeric vector representing the term or vector.}

\item{matrix}{A numeric matrix or a sparse matrix against which the similarity is calculated.}

\item{n}{Number of nearest neighbors to return.}
}
\value{
A data frame with the ranks, terms, and their cosine similarity scores.
}
\description{
This function identifies the nearest neighbors of a given term or vector
in a matrix based on cosine similarity.
}
