% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_scrape_urls.R
\name{standardize_date}
\alias{standardize_date}
\title{Standardize Date Format}
\usage{
standardize_date(date_str)
}
\arguments{
\item{date_str}{A character string representing a date.}
}
\value{
A character string representing the standardized date in "YYYY-MM-DD" format, or NA if the date cannot be parsed.
}
\description{
This function attempts to parse a date string using multiple formats and
standardizes it to "YYYY-MM-DD". It first tries ISO 8601 formats,
and then common formats like ymd, dmy, and mdy.
}
