% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_write_jsonl.R
\name{x_write_jsonl}
\alias{x_write_jsonl}
\title{Write DataFrame to a JSON Lines File}
\usage{
x_write_jsonl(df, file_name)
}
\arguments{
\item{df}{The data frame to be written to file.}

\item{file_name}{The path to the file where the data frame should be written.}
}
\description{
This function takes a data frame and writes it to a file in the JSON Lines format,
where each line is a JSON representation of one row of the data frame.
}
\examples{
# Example usage
# df <- data.frame(a = 1:3, b = letters[1:3])
# nlp_write_jsonl(df, "example.jsonl")
}
