% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbreviations.R
\docType{data}
\name{abbreviations}
\alias{abbreviations}
\title{Common Abbreviations for Sentence Splitting}
\format{
A character vector with some common English abbreviations.
}
\source{
Developed internally for sentence splitting functionality.
}
\usage{
abbreviations
}
\description{
A character vector of common abbreviations used in English.
These abbreviations are used to assist in sentence splitting,
ensuring that sentence boundaries are not incorrectly identified
at these abbreviations.
}
\keyword{datasets}
