% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp_cast_tokens.R
\name{nlp_cast_tokens}
\alias{nlp_cast_tokens}
\title{Convert token list to data frame}
\usage{
nlp_cast_tokens(tok)
}
\arguments{
\item{tok}{List with at least a \code{tokens} element (and optionally \code{spans}), e.g. output of \code{nlp_tokenize_text(..., include_spans = TRUE)}.}
}
\value{
Data frame with columns for unit id, token, and optionally start/end spans.
}
\description{
Convert the token list returned by \code{\link{nlp_tokenize_text}} into a data
frame (long format), with identifiers and optional spans.
}
\examples{
tok <- list(
  tokens = list(
    "1.1" = c("Hello", "world", "."),
    "1.2" = c("This", "is", "an", "example", "."),
    "2.1" = c("This", "is", "a", "party", "!")
  )
)
dtm <- nlp_cast_tokens(tok)

}
