% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp_index_tokens.R
\name{nlp_index_tokens}
\alias{nlp_index_tokens}
\title{Create a BM25 Search Index}
\usage{
nlp_index_tokens(tokens, k1 = 1.2, b = 0.75, stem = FALSE)
}
\arguments{
\item{tokens}{A named list of character vectors.}

\item{k1}{BM25 saturation parameter.}

\item{b}{BM25 length normalization.}

\item{stem}{Logical; if TRUE, stems tokens.}
}
\description{
Create a BM25 Search Index
}
