% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp_tokenize_text.R
\name{nlp_tokenize_text}
\alias{nlp_tokenize_text}
\title{Tokenize Text Data (mostly) Non-Destructively}
\usage{
nlp_tokenize_text(tif, by = "text_id")
}
\arguments{
\item{tif}{A data frame containing the text to be tokenized and a document identifier in 'doc_id'.}

\item{by}{A character string specifying grouping column.}
}
\value{
A named list of tokens, where each list item corresponds to a document.
}
\description{
This function tokenizes text data from a data frame using the 'tokenizers' package, preserving the original text structure like capitalization and punctuation.
}
