% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_corpus.R
\name{search_corpus}
\alias{search_corpus}
\title{NLP Search Corpus}
\usage{
search_corpus(
  tif,
  search,
  n = 0,
  is_inline = FALSE,
  highlight = c("<b>", "</b>"),
  cores = 1
)
}
\arguments{
\item{tif}{A data frame or data.table containing the text corpus.}

\item{search}{The search pattern or query.}

\item{n}{Numeric, default 0. Specifies the context size around the found patterns.}

\item{is_inline}{Logical, default FALSE. Indicates if the search should be inline.}

\item{highlight}{A character vector of length two, default c('\if{html}{\out{<b>}}', '\if{html}{\out{</b>}}').
Used to highlight the found patterns in the text.}

\item{cores}{Numeric, default 1. The number of cores to use for parallel processing.}
}
\value{
A data.table with the search results.
}
\description{
Searches a text corpus for specified patterns, with support for parallel processing.
}
