% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_index.R
\name{search_index}
\alias{search_index}
\title{Search the BM25 Index}
\usage{
search_index(index, query, n = 10, stem = FALSE)
}
\arguments{
\item{index}{A data.table created by nlp_index_tokens.}

\item{query}{A character string.}

\item{n}{Number of results to return.}

\item{stem}{Logical; must match the setting used during indexing.}
}
\value{
A data.table of results ranked by score.
}
\description{
Search the BM25 Index
}
