% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textpress-package.R
\docType{package}
\name{textpress-package}
\alias{textpress}
\alias{textpress-package}
\title{textpress: A Lightweight and Versatile NLP Toolkit}
\description{
A simple Natural Language Processing (NLP) toolkit focused on search-centric workflows with minimal dependencies. The package offers key features for web scraping, text processing, corpus search, and text embedding generation via the 'HuggingFace API' \url{https://huggingface.co/docs/api-inference/index}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jaytimm/textpress}
  \item \url{https://jaytimm.github.io/textpress/}
  \item Report bugs at \url{https://github.com/jaytimm/textpress/issues}
}

}
\author{
\strong{Maintainer}: Jason Timm \email{JaTimm@salud.unm.edu}

}
\keyword{internal}
