% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textpress-package.R
\docType{package}
\name{textpress-package}
\alias{textpress}
\alias{textpress-package}
\title{textpress: A lightweight, unobtrusive R companion for basic NLP.}
\description{
A lightweight, versatile NLP companion in R. The package provides features for (1) basic text processing, (2) corpus search, and (3) web scraping. Additionally included are utility functions for (4) building text embeddings via the HuggingFace API, and (5) fetching chat completions via the OpenAI API.
}
\keyword{internal}
