% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textpress-package.R
\docType{package}
\name{textpress-package}
\alias{textpress}
\alias{textpress-package}
\title{textpress: A lightweight and versatile NLP toolkit for R}
\description{
A simple NLP toolkit focused on search-centric workflows with minimal dependencies. The package offers key features for web scraping, text processing, corpus search, and text embedding generation via the HuggingFace API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jaytimm/textpress}
  \item \url{https://jaytimm.github.io/textpress/}
  \item Report bugs at \url{https://github.com/jaytimm/textpress/issues}
}

}
\keyword{internal}
